+++ sdk/textanalytics/azure-ai-textanalytics/src/samples/java/com/azure/ai/textanalytics/lro/AnalyzeHealthcareEntities.java
@@ -74,8 +74,8 @@ public class AnalyzeHealthcareEntities {
                 resultCollection.getModelVersion());
             // Batch statistics
             TextDocumentBatchStatistics batchStatistics = resultCollection.getStatistics();
-            System.out.printf("Documents statistics: document count = %s, erroneous document count = %s, "
-                                  + "transaction count = %s, valid document count = %s.%n",
+            System.out.printf("Documents statistics: document count = %d, erroneous document count = %d, "
+                                  + "transaction count = %d, valid document count = %d.%n",
                 batchStatistics.getDocumentCount(), batchStatistics.getInvalidDocumentCount(),
                 batchStatistics.getTransactionCount(), batchStatistics.getValidDocumentCount());